test: split oversized auth test files under the length budget
AuthCoordinatorStaleRevalidationTests keeps the sign-out-vs-in-flight
races; the stale-clear-vs-fresh-sign-in suite moves to
AuthCoordinatorStaleClearTests. HostBrowserSignInFlowTests' private
fakes move to HostBrowserSignInFlowFakes. No test behavior changes.
